refactor actions into multiple modules
